% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_3d}
\alias{aggregate_3d}
\title{Aggregate a 3D Array Along Specified Dimensions}
\usage{
aggregate_3d(ary_Data, n_Aggregate, fct_Aggregate = mean, ...)
}
\arguments{
\item{ary_Data}{3D (time, x, y) numeric array to be aggregated. The first must be time dimension.}

\item{n_Aggregate}{An integer specifying the factor by which to aggregate the array along dimensions 2 and 3.}

\item{fct_Aggregate}{A function to be applied to each aggregated block (default is \code{mean}).}

\item{...}{Additional arguments to be passed to \code{fct_Aggregate}.}
}
\value{
A 3D numeric array with reduced dimensions after aggregation.
}
\description{
Aggregates a 3D array by reducing its resolution based on a specified aggregation factor.
}
