% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale_temperature_UNF.R
\name{downscale_temperature_UNF}
\alias{downscale_temperature_UNF}
\title{Downscale temperature using UNF data}
\usage{
downscale_temperature_UNF(
  idx_Continent,
  fn_GCRC_Ref,
  fn_Temp_Ref,
  fn_Write,
  num_Factor = 100
)
}
\arguments{
\item{idx_Continent}{(string) Index of continents from ("eu", "af","as","au","na","sa")}

\item{fn_GCRC_Ref}{(string) File path to GCRC of original temprature data}

\item{fn_Temp_Ref}{(string) File path to temperature data}

\item{fn_Write}{(string) File path to write downscaled temperature data}

\item{num_Factor}{(num) Scaling factor for temperature correction}
}
\description{
Downscale temperature using UNF data
}
