% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcrc.R
\name{find_gcrc_location}
\alias{find_gcrc_location}
\title{Find a specific GCRC location within a continent}
\usage{
find_gcrc_location(str_Continent, id_GCRC)
}
\arguments{
\item{str_Continent}{string, It must be one of "eu" (Europe), "af" (Africa), "as" (Asia), "au" (Australia), "na" (North America), or "sa" (South America).}

\item{id_GCRC}{integer, ID of the GCRC within the specified continent}
}
\value{
A data frame containing the information of the specified GCRC location.
}
\description{
This function retrieves the information of a specific GCRC location within a given continent.
}
\examples{
# Example usage:
# Assuming lst_GCRC is a list containing information of GCRCs indexed by continent,
# and we want to retrieve the information of GCRC with ID 1 in continent "eu" (Europe):
# find_gcrc_location("eu", 1)
}
