% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_direction.R
\name{get_inflow_cells}
\alias{get_inflow_cells}
\title{Some tools for flow direction / path / net}
\usage{
get_inflow_cells(num_FlowDir)
}
\arguments{
\item{num_FlowDir}{Integer. The cell number of the next cell. The cell number must range from 1 to the length of the cells.
If the cell has no outflow, the number should be set to itself.}
}
\value{
A list of integers representing the cells that flow into the specified cell.
}
\description{
This package provides some tools for handling flow direction, path, and network-related calculations.
}
\section{Functions}{

\itemize{
\item \code{get_inflow_cells}: Get all the cells which will flow into a specified cell (including the cell itself).
}
}

\examples{
num_FD <- c(4,5,2,5,9,3,7,8,8)
lst_Cells <- get_inflow_cells(num_FD)
}
