% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflow_WGAP3.R
\name{inflow_cell_GCRC}
\alias{inflow_cell_GCRC}
\title{Get the inflow Cells with GCRC Number}
\usage{
inflow_cell_GCRC(idx_Continent, num_GCRC)
}
\arguments{
\item{idx_Continent}{(string) Index of continents from ("eu", "af","as","au","na","sa")}

\item{num_GCRC}{(int) GCRC number}
}
\value{
(vector of int) Inflow cells of every this WaterGAP3 Cell
}
\description{
Get the inflow Cells with GCRC Number
}
