% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{linear_interpolate}
\alias{linear_interpolate}
\title{Linear Interpolation Between Two Vectors}
\usage{
linear_interpolate(num_Start, num_End, n_Inter)
}
\arguments{
\item{num_Start}{A numeric vector representing the start points.}

\item{num_End}{A numeric vector representing the end points.}

\item{n_Inter}{An integer specifying the number of interpolation points.}
}
\value{
A matrix where each row corresponds to an interpolated point
between \code{num_Start} and \code{num_End}.
}
\description{
This function performs linear interpolation between two vectors
\code{num_Start} and \code{num_End} and returns the interpolated results in a matrix.
}
\details{
The function creates \code{n_Inter} equally spaced interpolations
between the corresponding elements of \code{num_Start} and \code{num_End}.
}
\examples{
num_Start <- c(1, 5, 10)
num_End <- c(10, 15, 20)
n_Inter <- 5
interpolation_matrix <- linear_interpolate(num_Start, num_End, n_Inter)

}
