% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{aggregate_wg3}
\alias{aggregate_wg3}
\title{Aggregate Raster Data}
\usage{
aggregate_wg3(rast_Data, int_Fact = 10, fct_Aggre = mean, fill_NA = 0)
}
\arguments{
\item{rast_Data}{A raster object containing the data to be aggregated.}

\item{int_Fact}{An integer specifying the aggregation factor. Default is 10.}

\item{fct_Aggre}{A function to be used for aggregation. Default is \code{mean}.}

\item{fill_NA}{A value to fill in for NA values in the aggregated raster.
Default is 0.}
}
\value{
A raster object of the aggregated data with NA values replaced.
}
\description{
This function aggregates raster data using a specified factor and function,
and fills NA values with a specified value.
}
