% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin_WGAP3.R
\name{basin_GCRC}
\alias{basin_GCRC}
\title{Get the basin of one Cell with GCRC Number}
\usage{
basin_GCRC(idx_Continent, num_GCRC)
}
\arguments{
\item{idx_Continent}{(string) Index of continents from ("eu", "af","as","au","na","sa")}

\item{num_GCRC}{(int) GCRC number}
}
\description{
Get the basin of one Cell with GCRC Number
}
