% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unf.R
\name{unf}
\alias{unf}
\alias{read_UNF}
\alias{read_UNF_scp}
\alias{write_UNF}
\alias{unf_2_raster}
\alias{unf_2_raster_merge}
\alias{raster_2_unf}
\alias{extract_unf}
\title{Data im- & export in UNF file}
\usage{
read_UNF(fn_UNF)

read_UNF_scp(fn_UNF, ip_Host, str_Username, str_Password)

write_UNF(data_Export, fn_UNF)

unf_2_raster(idx_Continent, num_Data)

unf_2_raster_merge(lst_num_Data)

raster_2_unf(rast_Data, idx_Continent)

extract_unf(rast_Data, idx_Continent, fct_Extract = mean, fill_NA = 0)
}
\arguments{
\item{fn_UNF}{string, name of the UNF file}

\item{ip_Host}{string, the name of the remote host or its IP address}

\item{str_Username}{string, the name of the user on the remote machine}

\item{str_Password}{string, a password for accessing the local SSH key}

\item{data_Export}{vector, data, that would exported}

\item{idx_Continent}{string in ("eu", "af", "as", "au", "na", "sa", "global_wg2")}

\item{num_Data}{vector (1D), data, that read from UNF-file and order by the GCRC-Number}

\item{lst_num_Data}{list of vector, data, that read from UNF-file and order by the GCRC-Number,
the list must named as ("eu", "af", "as", "au", "na", "sa") two or more}

\item{rast_Data}{(terra::SpatRaster) global raster data and the resolution must be 5'}

\item{fct_Extract}{A function to apply to the extracted values. Default is \code{mean}.}
}
\value{
vector of numric or integer by read, UNF file by write
}
\description{
Data im- & export in UNF file
}
\section{UNF Format}{
\itemize{
\item UNF0 / UNF8: \strong{numeric} with 4 / 8 Bytes,
\item UNF1, UNF2 / UNF4: \strong{integer} with 1, 2, / 4 Bytes
}
}

\section{Read & Write}{
\itemize{
\item Read data from UNF file
\item Write data to UNF file
}
}

