% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unf.R
\name{unf}
\alias{unf}
\alias{read_UNF}
\alias{read_UNF_scp}
\alias{write_UNF}
\title{Data im- & export in UNF file}
\usage{
read_UNF(fn_UNF)

read_UNF_scp(fn_UNF, ip_Host, str_Username, str_Password)

write_UNF(data_Export, fn_UNF)
}
\arguments{
\item{fn_UNF}{string, name of the UNF file}

\item{ip_Host}{string, the name of the remote host or its IP address}

\item{str_Username}{string, the name of the user on the remote machine}

\item{str_Password}{string, a password for accessing the local SSH key}

\item{data_Export}{vector, data, that would exported}
}
\value{
vector of numric or integer by read, UNF file by write
}
\description{
Data im- & export in UNF file
}
\section{UNF Format}{
\itemize{
\item UNF0 / UNF8: \strong{numeric} with 4 / 8 Bytes,
\item UNF1, UNF2 / UNF4: \strong{integer} with 1, 2, / 4 Bytes
}
}

\section{Read & Write}{
\itemize{
\item Read data from UNF file
\item Write data to UNF file
}
}

