% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{write_nc_global}
\alias{write_nc_global}
\title{Write 3D Array Data to a NetCDF File}
\usage{
write_nc_global(
  ary_Global,
  fn_NC,
  num_TimeDim,
  unit_TimeDim,
  name_Data,
  unit_Data,
  longname_Data,
  num_Resolut = 5
)
}
\arguments{
\item{ary_Global}{array (time, lon, lat),  from unf_2_ary_WG3(), default 5' (4320 * 2160)}

\item{fn_NC}{str, filename}

\item{num_TimeDim}{numric, vector of time dimension}

\item{unit_TimeDim}{str, unit of time dimension, like "year", "day from 1990-01-01"}

\item{name_Data, unit_Data, longname_Data}{str, data name, longname and unit for nc-file}

\item{num_Resolut}{int, resolution in arc minute}
}
\description{
Saves a 3D array (time, longitude, latitude) to a NetCDF file with appropriate metadata.
}
